Add option to backup and prune for tags
This adds the options for both backup and prune to allow specifying tags. For backup it adds the specified tags to the backup snapshot. For prune it will ensure only snapshots with the specified tags are forgotten.

Tested by adding unit tests and then:
```
poetry run pytest -k test_runner
```